api/src/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
distro==1.9.0
Django==4.2.9
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
insultgenerator==1.0.3
kombu==5.3.4
loremipsum==1.0.5
Markdown==3.5.1
pydantic==2.4.2
openai==1.6.1
packaging==23.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic_core==2.10.1
pydub==0.25.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tiktoken
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
edge-tts==6.1.10
python-docx==1.1.2
pypdf2==3.0.1
paycomuz==3.1
payze-pkg
groq
